Reformated Nav bar to become a component
